Fixed database configurations

+++ src/main/java/br/com/vinibelo/githubprofilemanagermutant/controllers/dto/users/UsersToListDto.java
@@ -1,10 +1,9 @@
 package br.com.vinibelo.githubprofilemanagermutant.controllers.dto.users;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public record UsersToListDto(
-        BigInteger id,
+        Long id,
         String login,
         String url,
         List<RoleDto> roles
