+++ src/main/java/br/com/vinibelo/githubprofilemanagermutant/repositories/ProfileRepository.java
@@ -4,7 +4,5 @@ import br.com.vinibelo.githubprofilemanagermutant.entities.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface ProfileRepository extends JpaRepository<Profile, BigInteger> { }
+public interface ProfileRepository extends JpaRepository<Profile, Long> { }
