+++ src/main/java/br/com/vinibelo/githubprofilemanagermutant/repositories/UserRepository.java
@@ -4,9 +4,7 @@ import br.com.vinibelo.githubprofilemanagermutant.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
     boolean existsUserByLogin(String login);
 }
