+++ src/main/java/br/com/vinibelo/githubprofilemanagermutant/services/ProfilesService.java
@@ -4,7 +4,6 @@ import br.com.vinibelo.githubprofilemanagermutant.entities.Profile;
 import br.com.vinibelo.githubprofilemanagermutant.repositories.ProfileRepository;
 import org.springframework.stereotype.Service;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Service
@@ -15,7 +14,7 @@ public class ProfilesService {
         this.profileRepository = profileRepository;
     }
 
-    public void createProfile(String name, String url, Optional<BigInteger> userId) {
+    public void createProfile(String name, String url, Optional<Long> userId) {
         Profile profile = new Profile(name, url);
         profileRepository.save(profile);
     }
